Renamed solution files. Suppressed SDK warnings
