+++ Visualization/Web/Dockerfile
@@ -1,0 +1,19 @@
+FROM mcr.microsoft.com/dotnet/core/aspnet:3.0-stretch-slim AS base
+WORKDIR /app
+EXPOSE 80
+
+FROM mcr.microsoft.com/dotnet/core/sdk:3.0-stretch AS build
+WORKDIR /src
+COPY ["Web/Web.csproj", "Web/"]
+RUN dotnet restore "Web/Web.csproj"
+COPY . .
+WORKDIR "/src/Web"
+RUN dotnet build "Web.csproj" -c Release -o /app
+
+FROM build AS publish
+RUN dotnet publish "Web.csproj" -c Release -o /app
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app .
+ENTRYPOINT ["dotnet", "Web.dll"]
