+++ Visualization/Web/Startup.cs
@@ -9,7 +9,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Inzynierka
+namespace Web
 {
     public class Startup
     {
