+++ Visualization/Web/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-﻿@using Inzynierka
-@using Inzynierka.Models
+﻿@using Web
+@using Web.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
